TEST: Adjusted tests which make sure that Opener uses GzipFile/IndexedGzipFile classes as appropriate
